+++ package.json
@@ -33,7 +33,9 @@
 		"builder:iter": ".venv/bin/python builder/core/cli.py iter:run",
 		"test": "vitest run",
 		"test:watch": "vitest",
-		"test:python": ".venv/bin/python -m pytest builder/tests/test_golden_checks.py -v",
+		"test:python": ".venv/bin/python -m pytest tests/integration/test_golden_checks.py -v",
+		"test:unit": ".venv/bin/python -m pytest tests/unit/ -v",
+		"test:integration": ".venv/bin/python -m pytest tests/integration/ -v",
 		"test:all": "npm run test && npm run test:python",
 		"build": "tsc -p tsconfig.build.json",
 		"spec:hello": ".venv/bin/python builder/utils/gen_spec.py",
