+++ tests/__init__.py
@@ -1,0 +1,9 @@
+"""
+Test suite for Code Builder.
+
+This package contains all tests for the Code Builder system, organized by type:
+- unit/ - Unit tests for individual components
+- integration/ - Integration tests and golden checks
+- data/ - Test data files
+- scripts/ - Test scripts and workflows
+"""
