+++ tests/integration/__init__.py
@@ -1,0 +1,6 @@
+"""
+Integration tests and golden checks for Code Builder.
+
+These tests verify the system works correctly as a whole and maintain
+compatibility as the system grows.
+"""
