+++ tests/unit/__init__.py
@@ -1,0 +1,5 @@
+"""
+Unit tests for Code Builder components.
+
+These tests verify individual components work correctly in isolation.
+"""
